fix: i64で座標計算すると誤差が出るのでとりあえずf64で座標を保持する。
後で整数に直したい

+++ src/types.rs
@@ -2,12 +2,12 @@
 pub struct Point {
     pub id: u32,
     pub name: String,
-    pub x: i32,
-    pub y: i32
+    pub x: f64,
+    pub y: f64
 }
 
 impl Point {
-    pub fn new(id: u32, name: String, x: i32, y: i32) -> Self {
+    pub fn new(id: u32, name: String, x: f64, y: f64) -> Self {
         Point { id, name, x, y, }
     }
 }
